relocate align_up to public headers
